# lib/api-zod/src/generated/types/uploadUrlInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UploadUrlInputContentType } from './uploadUrlInputContentType';

export interface UploadUrlInput {
  name: string;
  /** @minimum 1 */
  size: number;
  contentType: UploadUrlInputContentType;
}